refactor characteristics_util
